# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CheeseHacks API",
    "description": "Development documentation for CheeseHacks that is automatically generated on server start. See swagger-autogen for how to document endpoints when writing API calls.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/team-utility/team/": {
      "get": {
        "tags": [
          "team-utility/team"
        ],
        "summary": "Load all teams",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "team-utility/team"
        ],
        "summary": "Create a team",
        "description": "Uses information provided to create a new team.",
        "parameters": [
          {
            "name": "team",
            "in": "body",
            "description": "Creating new team",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "The Best Team"
                },
                "description": {
                  "type": "string",
                  "example": "We want to use javascript to win the biggest prize!"
                },
                "team_captain": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Successfully created new team."
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/team-utility/team/{id}": {
      "get": {
        "tags": [
          "team-utility/team"
        ],
        "summary": "Load a single team",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "team-utility/team"
        ],
        "summary": "Modify a team",
        "description": "Allows the team captain of a team to modify their team. Admins can modify any team",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "team",
            "in": "body",
            "description": "Creating new team",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "The Best Team"
                },
                "description": {
                  "type": "string",
                  "example": "We want to use javascript to win the biggest prize!"
                },
                "team_captain": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/team-utility/team/judging/{id}": {
      "get": {
        "tags": [
          "team-utility/team"
        ],
        "summary": "Load a single team with judging details",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}